docs: update KLIP-26 with latest interfaces (#5472)
