+++ ksqldb-api-client/src/main/java/io/confluent/ksql/api/client/exception/KsqlException.java
@@ -13,15 +13,17 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.ksql.api.client;
+package io.confluent.ksql.api.client.exception;
 
-public class KsqlClientException extends RuntimeException {
+/**
+ * Converts {@code KsqlErrorMessage} into an exception to be passed to client exception handlers.
+ */
+public class KsqlException extends RuntimeException {
 
-  public KsqlClientException(final String message) {
+  private final int errorCode;
+
+  public KsqlException(final int errorCode, final String message) {
     super(message);
-  }
-
-  public KsqlClientException(final String message, final Throwable throwable) {
-    super(message, throwable);
+    this.errorCode = errorCode;
   }
 }
